requirements.txt
numpy==1.19.5
python-dotenv==0.19.1
python_telegram_bot==13.1
tensorflow==2.4.1
nltk==3.5